Move get_kos_clash_values() to npc_combat and write test
