+++ npc_combat.py
@@ -34,3 +34,10 @@ def get_mod_message(mod_value):
     if mod_value == "Star":
         return "Defense is reversal."
     return "No modification."
+
+
+def get_kos_clash_values():
+    kos_atk = get_attack_value(random.randint(1, 6))
+    kos_blk = get_blk_value(random.randint(1, 6))
+    kos_mod = get_mod_value(random.randint(1, 6))
+    return kos_atk, kos_blk, kos_mod
